Explicit resources freeing
It improves resource management. For example I try to build a gfm example with de_window library instead of sdl2, de_window doesn't allow to allocate a window on stack and allocating a window on heap causes segfaults during garbage collection (in GLShader dtor). This PR prevents segfault, because shaders are destroyed explicitly.